requirements.txt
Pillow==11.1.0
requests==2.32.3
numpy==2.2.3
